// frontend/src/generated/models/ModelDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ModelDto = {
    id?: string;
    name?: string | null;
    type?: string | null;
    provider?: string | null;
    baseUrl?: string | null;
    hfRepo?: string | null;
    modelFile?: string | null;
    downloaded?: boolean | null;
    downloadStatus?: string | null;
    lastUsedAt?: string | null;
    apiKey?: string | null;
    hasHfToken?: boolean;
    createdAt?: string;
    updatedAt?: string;
};